backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nebula_cloud"
version = "1.0.0"
description = "Бэкенд приложение для облачного хранилища данных на Django"
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = { text = "MIT" }
authors = [{ name = "Sergey Tormozov", email = "sergei.tormozow@yandex.ru" }]

dependencies = [
    "Django>=6.0.0,<6.1.0",
    "djangorestframework>=3.15.0,<3.16.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-environ>=0.12.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "nanoid>=2.0.0",
    "django-cors-headers>=4.9.0",
    "drf-spectacular>=0.29.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.4.0",
    "isort>=5.13.0",
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "django-debug-toolbar>=4.0.0",
    "factory-boy>=3.3.0",
    "faker>=40.4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = [
    "media",
    "staticfiles",
    "static",
    "node_modules",
    ".venv",
    "venv",
    "__pycache__",
    "migrations",
    "*.migrations",
    "tests",
    "*.tests",
]

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.vscode
  | migrations
  | node_modules
  | __pycache__
  | static
  | media
)/
'''

[tool.ruff]
line-length = 100
target-version = "py312"

# Новые настройки в секции lint
[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (black handles this)
    "E203", # whitespace before ':'
]
extend-fixable = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["backend", "users", "storage"]

[tool.ruff.lint.per-file-ignores]
"**/migrations/*.py" = ["ALL"]
"**/tests/**/*.py" = ["ALL"]
"test_*.py" = ["ALL"]
"conftest.py" = ["ALL"]
"**/settings.py" = ["F401", "F403"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-fail-under=80",
    "--strict-markers",
    "--tb=short",
    "-v",
]

[tool.coverage.run]
branch = true
source = ["storage", "users"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
    "*/.venv/*",
    "*/manage.py",
    "*/asgi.py",
    "*/wsgi.py",
    "*/apps.py",
    "*/admin.py",
    "*/core/settings.py",
    "*/core/urls.py",
    "*/core/logging_config.py",
    "*/storage/logging.py",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if settings.DEBUG",
    "logger\\.",
    '"""',
    "'''",
]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"
show_contexts = true

[tool.coverage.xml]
output = "logs/coverage.xml"
